Added RevitTorsionTools.msi
The RevitTorsionTools.msi will install the Torsion Tools toolbar in Revit 2020. This will enable folks to leverage the toolbar without having to open and compile the VS Solution. Please install and test the toolbar prior to using in any production environment, and that Torsion Tools is not liable for any real or digital loss from the installation or use of this installer and/or application. This application does not collect, store or transmit any data to any parties including Torsion Tools.